% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommonTS.R
\name{getCommonTS_DLY}
\alias{getCommonTS_DLY}
\title{Get Common Time series}
\usage{
getCommonTS_DLY(pxList)
}
\description{
The function provides the longest possible common time series for all
securities/indices in the price list pxList.
The function output is a list with two data.frame's. The TS data.frame holds the indexed time series
and the FD data.frame the first log differences.
getCommonTS_DLY()
}
